Use grub2-mkconfig for writing grub config
Instead of creating a template based grub config file, use
grub2-mkconfig to create the config file during the build
process.

This is a POC that just overwrites the kiwi written grub
config with a call of grub2-mkconfig during build. It seems
to work but I can't tell if it works generally. Also the
refactoring if continued has to write config file additions
like timeout, console setup and more as extra files to
/etc/grub.d/

Issue #1194

!!! WIP don't merge !!!